core/ansible/ansible.cfg: set 'become_method' to 'sudo'
This is an artifact of me copying and pasting from
jcmdln/openbsd-ansible and not running into any issues with 'become'
despite using the wrong value. Did this work because it automatically
handles the fallback? I don't think that's what the documentation
states but whatever, let's set the correct value now.