Configure Vault to use Consul as a storage backend
Because Vault isn't yet configured to use tls, before interacting with
it you must run the following:

   $ export VAULT_ADDR='http://127.0.0.1:8200'

Once this is done, you can follow the official documentation and/or
guides on configuring Vault to proceed as desired. Eventually Vault
will be configured to use tls by default, though to be honest I don't
feel like working on it right this moment.