Flattening pass; Use system packaged ansible
It turns out that some of the setup instructions didn't quite work,
such as adding the official HashiCorp repo in such a way that it
excluded their vagrant* packages. Rather than detailing out even more
steps to perform, we ditch the requirement for setting up a virtualenv
and instead ship './ansible/setup.yml' which handles all the initial
work required to actually build. As part of these changes, I've also
eliminated some nesting after deciding that my eventual goal of
bundling './apps' in this repository is not a good idea. Hopefully this
makes things easier to follow in general.

The next bit of work to do is bifurcate starting services with '-dev'
so that we can work toward a more complex Vagrant example that runs a
sibling play (likely bundled with our existing './ansible' playbook) to
properly bootstrap a 3-node cluster. This should allow configuring one
of the instances as the initial cluster member, then handle any
authentication between all other instances. If this all goes well, then
we should have _most_ of the initial model for scaling out and we can
write some Terraform scripts to deploy to a commodity hosting provider.

This may or may not pan out, so expect more bulk commits like this one
in the near future.